requirements.txt
openai
speechrecognition
pyttsx3
pyaudio
sounddevice
scipy
fastapi
uvicorn
python-multipart
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
sqlite3